Organize coding chains under dedicated module
