+++ chain.patch
@@ -1,7 +1,7 @@
-diff --git a/server/agents/chains/IncrementalCodingChain.js b/server/agents/chains/IncrementalCodingChain.js
+diff --git a/server/agents/chains/coding/IncrementalCodingChain.js b/server/agents/chains/coding/IncrementalCodingChain.js
 index a103bd6213ffb757cfeb0f04502b11a59bbd52aa..3f0d1ad2e752eea13abae893552f4f9a561491d2 100644
---- a/server/agents/chains/IncrementalCodingChain.js
-+++ b/server/agents/chains/IncrementalCodingChain.js
+--- a/server/agents/chains/coding/IncrementalCodingChain.js
++++ b/server/agents/chains/coding/IncrementalCodingChain.js
 @@ -1,55 +1,35 @@
 -// Modular LangChain chain for ContextStepBuilderAgent
 -// Receives: { gameSource, plan, step }
@@ -79,8 +79,8 @@ index 6ed4160c2390321b2577cf6f4706a4204af040b2..3a2daa6405e37263a2b72b26906d248b
 --- a/server/tests/unit/IncrementalCodingChain.test.js
 +++ b/server/tests/unit/IncrementalCodingChain.test.js
 @@ -21,50 +21,61 @@ describe('IncrementalCodingChain', () => {
-     const systemPromptPath = path.join(__dirname, '../../agents/prompts/IncrementalCodingChain.system.prompt.md');
-     const humanPromptPath = path.join(__dirname, '../../agents/prompts/IncrementalCodingChain.human.prompt.md');
+     const systemPromptPath = path.join(__dirname, '../../agents/prompts/coding/IncrementalCodingChain.system.prompt.md');
+     const humanPromptPath = path.join(__dirname, '../../agents/prompts/coding/IncrementalCodingChain.human.prompt.md');
      systemPromptString = await fs.readFile(systemPromptPath, 'utf8');
      humanPromptString = await fs.readFile(humanPromptPath, 'utf8');
    });
