+++ server/agents/chains/coding/FeedbackChain.js
@@ -2,8 +2,8 @@
 // Receives: { runtimeLogs, stepId }
 // Returns: { retryTarget: 'fixer' | 'planner', suggestion: string }
 
-import { createJSONChain } from '../../utils/chainFactory.js';
-import { feedbackSchema } from '../../schemas/langchain-schemas.js';
+import { createJSONChain } from '../../../utils/chainFactory.js';
+import { feedbackSchema } from '../../../schemas/langchain-schemas.js';
 
 export const CHAIN_STATUS = {
   name: 'FeedbackChain',
@@ -16,7 +16,7 @@ export const CHAIN_STATUS = {
 async function createFeedbackChain(llm, options = {}) {
   return createJSONChain({
     chainName: 'FeedbackChain',
-    promptFile: 'FeedbackChain.prompt.md',
+    promptFile: 'coding/FeedbackChain.prompt.md',
     inputVariables: ['runtimeLogs', 'stepId'],
     schema: feedbackSchema,
     llm: llm, // Use provided LLM for backward compatibility
