+++ server/agents/chains/coding/IncrementalCodingChain.js
@@ -2,7 +2,7 @@
 // Receives: { gameSource, plan, step, entities }
 // Returns: revised JavaScript source as a string
 
-import { createChatChain } from '../../utils/chainFactory.js';
+import { createChatChain } from '../../../utils/chainFactory.js';
 
 export const CHAIN_STATUS = {
   name: 'IncrementalCodingChain',
@@ -15,8 +15,8 @@ export const CHAIN_STATUS = {
 export async function createIncrementalCodingChain(llm, options = {}) {
   const chain = await createChatChain({
     chainName: 'IncrementalCodingChain',
-    systemFile: 'IncrementalCodingChain.system.prompt.md',
-    humanFile: 'IncrementalCodingChain.human.prompt.md',
+    systemFile: 'coding/IncrementalCodingChain.system.prompt.md',
+    humanFile: 'coding/IncrementalCodingChain.human.prompt.md',
     inputVariables: ['gameSource', 'plan', 'step', 'entities'],
     preset: 'creative',
     llm,
