+++ server/agents/chains/design/EntityListBuilderChain.js
@@ -4,7 +4,7 @@ import { entityListBuilderSchema } from '../../../schemas/langchain-schemas.js';
 async function createEntityListBuilderChain(llm, options = {}) {
   return createJSONChain({
     chainName: 'EntityListBuilderChain',
-    promptFile: 'design/entity-list-builder.md',
+    promptFile: 'design/EntityListBuilderChain.prompt.md',
     inputVariables: ['context'],
     schema: entityListBuilderSchema,
     preset: 'structured',
