+++ server/agents/chains/design/IdeaGeneratorChain.js
@@ -4,7 +4,7 @@ import { ideaGeneratorSchema } from '../../../schemas/langchain-schemas.js';
 async function createIdeaGeneratorChain(llm, options = {}) {
   return createCreativeChain({
     chainName: 'IdeaGeneratorChain',
-    promptFile: 'design/idea-generator.md',
+    promptFile: 'design/IdeaGeneratorChain.prompt.md',
     inputVariables: ['context'],
     schema: ideaGeneratorSchema,
     llm: llm,
