+++ server/agents/chains/design/MechanicExtractorChain.js
@@ -4,7 +4,7 @@ import { mechanicExtractorSchema } from '../../../schemas/langchain-schemas.js';
 async function createMechanicExtractorChain(llm, options = {}) {
   return createJSONChain({
     chainName: 'MechanicExtractorChain',
-    promptFile: 'design/mechanic-extractor.md',
+    promptFile: 'design/MechanicExtractorChain.prompt.md',
     inputVariables: ['context'],
     schema: mechanicExtractorSchema,
     preset: 'structured',
