+++ server/agents/prompts/art/SpriteDesignChain.prompt.md
