+++ server/agents/prompts/coding/ControlBarTransformerAgent.prompt.md
