+++ server/agents/prompts/design/IdeaGeneratorChain.prompt.md
