+++ server/agents/prompts/design/MechanicExtractorChain.prompt.md
