+++ server/agents/prompts/design/WinConditionBuilderChain.prompt.md
